requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
lemon-markets-sdk==0.0.1
pytz==2020.1
requests==2.24.0
urllib3==1.25.11
